Add 1inch volume adapter
