+++ projects/1inch/api.js
@@ -1,0 +1,3 @@
+module.exports = {
+  endpoint: 'https://api.thegraph.com/subgraphs/name/1inch-exchange/oneinch-liquidity-protocol-v2',
+}
